# openapi/sesn.v1.yaml
openapi: 3.1.0
info:
  title: SES Navigator API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /proposals:
    post:
      operationId: createProposal
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProposalInput' }
      responses:
        '201':
          description: Draft proposal created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Proposal' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      operationId: listProposals
      parameters:
        - {
            name: q,
            in: query,
            schema: { type: string, minLength: 1, maxLength: 100 },
          }
        - {
            name: status,
            in: query,
            schema: { $ref: '#/components/schemas/ProposalStatus' },
          }
        - { name: cursor, in: query, schema: { type: string } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 200, default: 50 },
          }
      responses:
        '200':
          description: Proposal list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProposalList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /proposals/{id}:
    put:
      operationId: updateProposal
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: If-Match,
            in: header,
            required: true,
            schema: { type: string },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProposalInput' }
      responses:
        '200':
          description: Draft proposal updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Proposal' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
    get:
      operationId: getProposal
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        '200':
          description: Proposal detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Proposal' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /proposals/{id}/status:
    post:
      operationId: transitionProposalStatus
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: If-Match,
            in: header,
            required: true,
            schema: { type: string },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: '#/components/schemas/ProposalStatusTransitionInput' }
      responses:
        '200':
          description: Proposal status transitioned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Proposal' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /proposals/{id}/win:
    post:
      operationId: winProposal
      description: Atomically marks an offered proposal as won and creates an editable contract draft and initial engagement draft.
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: If-Match,
            in: header,
            required: true,
            schema: { type: string },
          }
        - {
            name: Idempotency-Key,
            in: header,
            required: true,
            schema:
              {
                type: string,
                minLength: 1,
                maxLength: 200,
                pattern: '^[A-Za-z0-9._:-]+$',
              },
          }
      responses:
        '200':
          description: Proposal won and draft identifiers returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProposalWinResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    ProposalStatus:
      type: string
      enum:
        [
          draft,
          pending_approval,
          approved,
          sent,
          interview_requested,
          interviewing,
          offered,
          won,
          lost,
          withdrawn,
          cancelled,
        ]
    Proposal:
      type: object
      required:
        [
          id,
          managementNo,
          projectPositionId,
          engineerId,
          destinationCompanyId,
          destinationContactId,
          resumeVersionId,
          requirementVersionId,
          proposedUnitPrice,
          currencyCode,
          status,
          proposedStartDate,
          validityDate,
          updatedAt,
          rowVersion,
        ]
      properties:
        id: { type: string, format: uuid }
        managementNo: { type: string }
        projectPositionId: { type: string, format: uuid }
        engineerId: { type: string, format: uuid }
        destinationCompanyId: { type: string, format: uuid }
        destinationContactId: { type: [string, 'null'], format: uuid }
        resumeVersionId: { type: [string, 'null'], format: uuid }
        requirementVersionId: { type: [string, 'null'], format: uuid }
        proposedUnitPrice: { type: [number, 'null'], minimum: 0 }
        currencyCode: { type: string, minLength: 3, maxLength: 3 }
        status: { $ref: '#/components/schemas/ProposalStatus' }
        proposedStartDate: { type: [string, 'null'], format: date }
        validityDate: { type: [string, 'null'], format: date }
        updatedAt: { type: string, format: date-time }
        rowVersion: { type: integer, minimum: 1 }
    ProposalInput:
      type: object
      required:
        [
          managementNo,
          projectPositionId,
          engineerId,
          destinationCompanyId,
          currencyCode,
        ]
      properties:
        managementNo: { type: string, minLength: 1, maxLength: 32 }
        projectPositionId: { type: string, format: uuid }
        engineerId: { type: string, format: uuid }
        destinationCompanyId: { type: string, format: uuid }
        destinationContactId: { type: [string, 'null'], format: uuid }
        resumeVersionId: { type: [string, 'null'], format: uuid }
        requirementVersionId: { type: [string, 'null'], format: uuid }
        proposedUnitPrice: { type: [number, 'null'], minimum: 0 }
        currencyCode:
          { type: string, minLength: 3, maxLength: 3, pattern: '^[A-Z]{3}$' }
        proposedStartDate: { type: [string, 'null'], format: date }
        validityDate: { type: [string, 'null'], format: date }
    ProposalStatusTransitionInput:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/ProposalStatus' }
        reason: { type: [string, 'null'], maxLength: 500 }
    ProposalWinResult:
      type: object
      additionalProperties: false
      required: [proposal, contractId, engagementId, created]
      properties:
        proposal: { $ref: '#/components/schemas/Proposal' }
        contractId: { type: string, format: uuid }
        engagementId: { type: string, format: uuid }
        created:
          type: boolean
          description: True when this request created drafts; false for an idempotent replay.
    ProposalList:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Proposal' } }
        page:
          type: object
          required: [limit, nextCursor]
          properties:
            limit: { type: integer }
            nextCursor: { type: ['string', 'null'] }
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
  responses:
    BadRequest:
      {
        description: Invalid request,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    Conflict:
      {
        description: Optimistic locking conflict,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    PreconditionRequired:
      {
        description: If-Match is required,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    Unauthorized:
      {
        description: Authentication required,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    Forbidden:
      {
        description: Permission denied,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    NotFound:
      {
        description: Not found or not visible,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }